rename pretty-text2 to pretty-text since there's only one
