+++ lib/default-config.js
@@ -24,7 +24,7 @@ import BooleanField from './components/fields/boolean';
 import PrettyBooleanField from './components/fields/pretty-boolean';
 import CheckboxBooleanField from './components/fields/checkbox-boolean';
 import CodeField from './components/fields/code';
-import PrettyTextField from './components/fields/pretty-text2';
+import PrettyTextField from './components/fields/pretty-text';
 import PrettyTagField from './components/helpers/pretty-tag';
 import ArrayField from './components/fields/array';
 import CheckboxArrayField from './components/fields/checkbox-array';
